+++ sources/Adapter/Bus.cpp
@@ -1,91 +1,136 @@
 #include "internal/Bus.hpp"
 #include "Burst.hpp"
+#include "ModbusTechnologyAdapter.hpp"
 
 namespace Technology_Adapter::Modbus {
 
-constexpr size_t NUM_READ_ATTEMPTS = 3; // 0 would mean instant failure
-
-Bus::Bus(Config::Bus const& config, Config::Portname const& actual_port)
-    : config_(config), actual_port_(actual_port), //
+Bus::Bus(ModbusTechnologyAdapter& owner, Config::Bus const& config,
+    Config::Portname const& actual_port,
+    // NOLINTNEXTLINE(modernize-pass-by-value)
+    Technology_Adapter::NonemptyDeviceRegistryPtr const& model_registry)
+    : owner_(owner), config_(config), actual_port_(actual_port), //
       logger_(HaSLI::LoggerManager::registerLogger(std::string(
           (std::string_view)("Modbus Bus " + config.id + "@" + actual_port)))),
-      context_(actual_port, config.baud, config.parity, config.data_bits,
+      model_registry_(model_registry),
+      connection_(actual_port, config.baud, config.parity, config.data_bits,
           config.stop_bits) {}
 
-void Bus::buildModel(
-    Information_Model::NonemptyDeviceBuilderInterfacePtr const& device_builder,
-    Technology_Adapter::NonemptyDeviceRegistryPtr const& model_registry) {
-
-  logger_->info("Registering all devices on bus {}", actual_port_);
-  std::vector<ConstString::ConstString> added;
-
+Bus::~Bus() {
   try {
-    try {
-      for (auto const& device : config_.devices) {
-        device_builder->buildDeviceBase(
-            std::string((std::string_view)device.id),
-            std::string((std::string_view)device.name),
-            std::string((std::string_view)device.description));
-        RegisterSet holding_registers(device.holding_registers);
-        RegisterSet input_registers(device.input_registers);
-        buildGroup(device_builder, model_registry, "", //
-            NonemptyPtr(shared_from_this()), //
-            device, holding_registers, input_registers, device);
-        if (model_registry->registrate(Information_Model::NonemptyDevicePtr(
-                device_builder->getResult()))) {
-
-          try {
-            added.push_back(device.id);
-          } catch (...) {
-            // `push_back` failed. This must be an out-of-memory.
-            model_registry->deregistrate(
-                std::string((std::string_view)device.id));
-            throw std::bad_alloc();
-          }
-        };
-      }
-    } catch (...) {
-      // Before re-throwing, deregister everything that has been registered.
-      for (auto const& device : added) {
-        model_registry->deregistrate(std::string((std::string_view)device));
-      }
-      throw;
-    }
-  } catch (std::exception const& exception) {
-    throw std::runtime_error(std::string((std::string_view)( //
-        "Deregistered all Modbus devices on bus " + actual_port_ +
-        " after: " + exception.what())));
+    stop();
   } catch (...) {
-    throw std::runtime_error(std::string((std::string_view)( //
-        "Deregistered all Modbus devices on bus " + actual_port_ +
-        " after a non-standard exception")));
   }
 }
 
-void Bus::start() { context_.lock()->connect(); }
+void Bus::buildModel(Information_Model::NonemptyDeviceBuilderInterfacePtr const&
+        device_builder) {
 
-void Bus::stop() { context_.lock()->close(); }
+  logger_->info("Registering all devices on bus {}", actual_port_);
+
+  auto accessor = connection_.lock();
+
+  try {
+    for (auto const& device : config_.devices) {
+      device_builder->buildDeviceBase(
+          std::string((std::string_view)device.id),
+          std::string((std::string_view)device.name),
+          std::string((std::string_view)device.description));
+      RegisterSet holding_registers(device.holding_registers);
+      RegisterSet input_registers(device.input_registers);
+      buildGroup(device_builder, "", //
+          NonemptyPtr(shared_from_this()), //
+          device, holding_registers, input_registers, device);
+      if (model_registry_->registrate(Information_Model::NonemptyDevicePtr(
+              device_builder->getResult()))) {
+
+        try {
+          accessor->registered_devices.push_back(device.id);
+        } catch (...) {
+          // `push_back` failed. This must be an out-of-memory.
+          model_registry_->deregistrate(
+              std::string((std::string_view)device.id));
+          throw std::bad_alloc();
+        }
+      } else {
+        abort(accessor,
+            "Deregistered all Modbus devices on bus " + actual_port_ +
+                " after registering " + device.id + "failed");
+      }
+    }
+  } catch (std::exception const& exception) {
+    abort(accessor,
+        "Deregistered all Modbus devices on bus " + actual_port_ +
+            " after: " + exception.what());
+  } catch (...) {
+    abort(accessor,
+        "Deregistered all Modbus devices on bus " + actual_port_ +
+            " after a non-standard exception");
+  }
+}
+
+void Bus::start() {
+  try {
+    auto accessor = connection_.lock();
+    accessor->context.connect();
+    accessor->connected = true;
+  } catch (...) {
+    stop();
+    throw;
+  }
+}
+
+void Bus::stop() {
+  auto accessor = connection_.lock();
+  stop(accessor);
+}
 
 // This has become too long to be a lambda. Hence a `Callable`
 struct Readcallback {
   Technology_Adapter::NonemptyDeviceRegistryPtr const model_registry;
   Bus::NonemptyPtr const bus;
-  int const slave_id;
   ConstString::ConstString const device_id;
-  std::shared_ptr<std::string> const metric_id; // to be initialized later
+  int const slave_id;
+  size_t const max_retries;
+  size_t const retry_delay;
+  std::shared_ptr<std::string> const metric_id; // initialized after constructor
   Config::Readable const readable;
   NonemptyPointer::NonemptyPtr<std::shared_ptr<BurstBuffer>> const buffer;
 
+  Readcallback(
+      // NOLINTBEGIN(readability-identifier-naming)
+      // NOLINTNEXTLINE(modernize-pass-by-value)
+      Technology_Adapter::NonemptyDeviceRegistryPtr const& model_registry_,
+      Bus::NonemptyPtr const& bus_, // NOLINT(modernize-pass-by-value)
+      Config::Device const& device, //
+      std::shared_ptr<std::string> metric_id_, //
+      Config::Readable readable_,
+      // NOLINTNEXTLINE(modernize-pass-by-value)
+      NonemptyPointer::NonemptyPtr<std::shared_ptr<BurstBuffer>> const& buffer_)
+      // NOLINTEND(readability-identifier-naming)
+      : model_registry(model_registry_), bus(bus_), device_id(device.id),
+        slave_id(device.slave_id), max_retries(device.max_retries),
+        retry_delay(device.retry_delay), metric_id(std::move(metric_id_)),
+        readable(std::move(readable_)), buffer(buffer_) {}
+
   Information_Model::DataVariant operator()() const {
     {
-      auto accessor = bus->context_.lock();
-      bus->logger_->debug("Reading {}", *metric_id);
-      accessor->setSlave(slave_id);
+      auto accessor = bus->connection_.lock();
+      if (accessor->connected) {
+        bus->logger_->debug("Reading {}", *metric_id);
+        accessor->context.setSlave(slave_id);
 
-      uint16_t* read_dest = buffer->padded.data();
-      for (auto const& burst : buffer->plan.bursts) {
-        readBurst(accessor, burst, read_dest);
-        read_dest += burst.num_registers;
+        uint16_t* read_dest = buffer->padded.data();
+        for (auto const& burst : buffer->plan.bursts) {
+          readBurst(accessor, burst, read_dest);
+          read_dest += burst.num_registers;
+        }
+      } else {
+        // Some other thread closed the connection. Hence the resource has been
+        // deregistered.
+        bus->logger_->debug(
+            "Reading {} failed because the connection was closed", *metric_id);
+        throw std::runtime_error(
+            (device_id + " has been deregistered").c_str());
       }
     } // no need to hold the lock during decoding
     size_t compact_size = buffer->compact.size();
@@ -97,8 +142,7 @@ struct Readcallback {
 
 private:
   void readBurst( //
-      Threadsafe::Resource<LibModbus::ContextRTU,
-          Threadsafe::QueuedMutex>::ScopedAccessor& accessor,
+      Bus::ConnectionResource::ScopedAccessor& accessor,
       BurstPlan::Burst const& burst, //
       uint16_t* read_dest) const {
 
@@ -115,50 +159,33 @@ private:
 
   // returns the number of registers actually read. That number is > 0
   int readRegisters( //
-      Threadsafe::Resource<LibModbus::ContextRTU,
-          Threadsafe::QueuedMutex>::ScopedAccessor& accessor,
+      Bus::ConnectionResource::ScopedAccessor& accessor,
       BurstPlan::Burst const& burst, //
       uint16_t* const read_dest, //
       RegisterIndex first_register, //
       int num) const {
     int num_read = 0;
-    size_t remaining_attempts = NUM_READ_ATTEMPTS;
+    size_t remaining_attempts = max_retries + 1;
     while ((num_read == 0) && (remaining_attempts > 0)) {
       try {
-        num_read =
-            accessor->readRegisters(first_register, burst.type, num, read_dest);
+        num_read = accessor->context.readRegisters(
+            first_register, burst.type, num, read_dest);
         if (num_read == 0) {
           bus->logger_->debug("Reading {} failed", *metric_id);
-          if (remaining_attempts > 1) {
-            bus->logger_->debug("Retrying to read {}", *metric_id);
-            // wait for next iteration
-          } else {
-            model_registry->deregistrate(
-                std::string((std::string_view)device_id));
-            throw std::runtime_error(std::string((std::string_view)( //
-                "Deregistered " + device_id + " after reading " + *metric_id +
-                " failed")));
-          }
+          retryOrAbort(remaining_attempts, accessor,
+              "Deregistered " + device_id +
+              " after too many read attempts for " + *metric_id);
         }
       } catch (LibModbus::ModbusError const& error) {
         bus->logger_->debug("Reading {} failed: {}", *metric_id, error.what());
         if (error.retryFeasible()) {
-          if (remaining_attempts > 1) {
-            bus->logger_->debug("Retrying to read {}", *metric_id);
-            // wait for next iteration
-          } else {
-            model_registry->deregistrate(
-                std::string((std::string_view)device_id));
-            throw std::runtime_error(std::string((std::string_view)( //
-                "Deregistered " + device_id +
-                " after too many read attempts. Last error was: " +
-                error.what())));
-          }
+          retryOrAbort(remaining_attempts, accessor,
+              "Deregistered " + device_id +
+              " after too many read attempts for " + *metric_id +
+              ". Last error was: " + error.what());
         } else {
-          model_registry->deregistrate(
-              std::string((std::string_view)device_id));
-          throw std::runtime_error(std::string((std::string_view)(
-              "Deregistered " + device_id + " after: " + error.what())));
+          bus->abort(accessor,
+              "Deregistered " + device_id + " after: " + error.what());
         }
       }
       --remaining_attempts;
@@ -166,19 +193,32 @@ private:
     // Now `num_read > 0`, because otherwise we have thrown
     return num_read;
   }
+
+  // retrying will happen after the function returns without throwing
+  void retryOrAbort( //
+      size_t& remaining_attempts,
+      Bus::ConnectionResource::ScopedAccessor& accessor,
+      ConstString::ConstString const& error_message) const {
+
+    if (remaining_attempts > 1) {
+      if (retry_delay > 0) {
+        std::this_thread::sleep_for(std::chrono::milliseconds(retry_delay));
+      }
+      bus->logger_->debug("Retrying to read {}", *metric_id);
+      // wait for next iteration of `ReadRegisters`
+    } else {
+      bus->abort(accessor, error_message);
+    }
+  }
 };
 
 void Bus::buildGroup(
     Information_Model::NonemptyDeviceBuilderInterfacePtr const& device_builder,
-    Technology_Adapter::NonemptyDeviceRegistryPtr const& model_registry,
     std::string const& group_id, //
     NonemptyPtr const& shared_this, //
     Config::Device const& device, //
     RegisterSet const& holding_registers, RegisterSet const& input_registers,
     Config::Group const& group) {
-
-  int slave_id = device.slave_id;
-  auto const& device_id = device.id;
 
   for (auto const& readable : group.readables) {
     auto buffer = NonemptyPointer::make_shared<BurstBuffer>( //
@@ -191,17 +231,38 @@ void Bus::buildGroup(
     *metric_id = device_builder->addReadableMetric( //
         group_id, std::string((std::string_view)readable.name),
         std::string((std::string_view)readable.description), readable.type,
-        Readcallback{model_registry, shared_this, slave_id, device_id,
-            metric_id, readable, buffer});
+        Readcallback(
+            model_registry_, shared_this, device, metric_id, readable, buffer));
   }
 
   for (auto const& subgroup : group.subgroups) {
     std::string group_id = device_builder->addDeviceElementGroup(
         std::string((std::string_view)subgroup.name),
         std::string((std::string_view)subgroup.description));
-    buildGroup(device_builder, model_registry, group_id, shared_this, device, //
+    buildGroup(device_builder, group_id, shared_this, device, //
         holding_registers, input_registers, subgroup);
   }
+}
+
+void Bus::stop(ConnectionResource::ScopedAccessor& accessor) {
+  logger_->trace("Stopping bus {}", actual_port_);
+  for (auto const& device : accessor->registered_devices) {
+    model_registry_->deregistrate(std::string((std::string_view)device));
+  }
+  accessor->registered_devices.clear();
+  if (accessor->connected) {
+    accessor->context.close();
+    accessor->connected = false;
+  }
+}
+
+void Bus::abort(ConnectionResource::ScopedAccessor& accessor,
+    ConstString::ConstString const& error_message) {
+
+  logger_->trace("Aborting bus {}", actual_port_);
+  stop(accessor);
+  owner_.cancelBus(actual_port_);
+  throw std::runtime_error(error_message.c_str());
 }
 
 } // namespace Technology_Adapter::Modbus