+++ sources/Adapter/Config.cpp
@@ -1,3 +1,5 @@
+#include <iostream>
+
 #include "Config.hpp"
 
 namespace Technology_Adapter::Modbus::Config {
@@ -20,13 +22,14 @@ Device::Device(ConstString::ConstString id_, ConstString::ConstString name,
     ConstString::ConstString description, std::vector<Readable> readables_,
     std::vector<Group> subgroups_,
     // NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
-    int slave_id_, size_t burst_size_,
+    int slave_id_, size_t burst_size_, size_t max_retries_, size_t retry_delay_,
     // NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
     std::vector<RegisterRange> const& holding_registers_,
     std::vector<RegisterRange> const& input_registers_)
     : Group(std::move(name), std::move(description), std::move(readables_),
           std::move(subgroups_)),
       id(std::move(id_)), slave_id(slave_id_), burst_size(burst_size_),
+      max_retries(max_retries_), retry_delay(retry_delay_),
       holding_registers(holding_registers_), input_registers(input_registers_) {
 }